Astrologers/requirements.txt
Pillow>=10.1.0
stripe>=5.0.0
requests>=2.28.0
